Change RouterMaster to RouterNexus and so on.
