+++ rzrtd/src/core/message/mod.rs
@@ -1,8 +1,8 @@
 //
 // ReZe.Rs - Router Daemon
-//   Copyright (C) 2018 Toshiaki Takada
+//   Copyright (C) 2018,2019 Toshiaki Takada
 //
 // Message
 //
-pub mod master;
+pub mod nexus;
 pub mod zebra;
