+++ rzrtd/src/core/mod.rs
@@ -1,9 +1,10 @@
 //
 // ReZe.Rs - Router Daemon
-//   Copyright (C) 2018 Toshiaki Takada
+//   Copyright (C) 2018,2019 Toshiaki Takada
 //
 // Core module.
 //
+pub mod nexus;
 pub mod master;
 pub mod protocols;
 pub mod message;
