+++ rzrtd/src/ospf/mod.rs
@@ -1,8 +1,9 @@
 //
 // ReZe.Rs - Router Daemon
-//   Copyright (C) 2018 Toshiaki Takada
+//   Copyright (C) 2018,2019 Toshiaki Takada
 //
 // OSPF Version 2
 //   Supported specifictions:
 //     RFC2328 TBD
+//
 pub mod master;
